Revolution: Get rid of NodeTypesOf!
NodeTypesOf allowed limiting the use of RankNTypes to KPointed,
but the ammount of complexity it added was too large.
Now KFunctor and KFoldable also use RankNTypes but everything is much simpler!

Note that in the new formulation, didn't succeed in making instances for Flip (ITerm a) e,
so it currently has its own combinator traverseITerm instead.