Move non-production / experimental modules to test
Most other modules are in actual use in Lamdu,
and not just in an example test language,
so their design is based on first-hand requirements.

An exception is the Diff module, which stays as an exported module
even though it wasn't necessary for any production use yet.
The reason is that it serves as a good compelling example
for the abilities of Hypertypes demonstrated in the README.